ping now works when multiple packets are received

+++ src/includes/context.h
@@ -7,17 +7,17 @@
 #include <netinet/ip_icmp.h>
 
 typedef struct s_ftp_ctx {
-	t_bool	verbose;
-	struct in_addr	ip;
-	char	*domain;
-	int		sock;
-	int		seq;
-	int		id;
-	unsigned char	*payload_ptr;
-	size_t	payload_size;
-	unsigned char	*packet_buffer;
-	size_t	packet_size;
-	size_t	response_size;
+	t_bool				verbose;
+	struct in_addr		ip;
+	char				*domain;
+	int					sock;
+	int					seq;
+	unsigned short		id;
+	unsigned char		*payload_ptr;
+	size_t				payload_size;
+	unsigned char		*packet_buffer;
+	size_t				packet_size;
+	size_t				response_size;
 	struct sockaddr_in	addr;
 } t_ftp_ctx;
